add pch

+++ build.sh
@@ -1,8 +1,12 @@
 #!/user/bin/env bash
 
-rm -rf build
-mkdir build
+if [ "$1" == "-r" ]; then
+    rm -rf build
+fi
+
+mkdir -p build
 cd build || exit 1
 cmake ..
 make || exit 1
 cd .. || exit 1
+
